+++ src/components/CreatePage/EmptyImage.jsx
@@ -2,21 +2,11 @@ import React from "react";
 
 const EmptyImage = () => {
   return (
-    <div className="w-full h-full max-w-[75rem] min-w-[0] max-h-[full] flex-1 overflow-hidden">
-      <div className="grid grid-cols-2 gap-[0.25rem] w-full h-full">
-        <div className="w-full h-full max-w-full rounded-3xl min-w-full bg-[#2D2D2D]">
-          이미지 1
-        </div>
-        <div className="w-full h-full max-w-full rounded-3xl min-w-full bg-[#2D2D2D]">
-          이미지 1
-        </div>
-        <div className="w-full h-full max-w-full rounded-3xl min-w-full bg-[#2D2D2D]">
-          이미지 1
-        </div>
-        <div className="w-full h-full max-w-full rounded-3xl min-w-full bg-[#2D2D2D]">
-          이미지 1
-        </div>
-      </div>
+    <div className="grid grid-cols-2 gap-1">
+      <div className="relative aspect-[5/3] w-full bg-[#2D2D2D]">이미지 1</div>
+      <div className="relative aspect-[5/3] w-full bg-[#2D2D2D]">이미지 1</div>
+      <div className="relative aspect-[5/3] w-full bg-[#2D2D2D]">이미지 1</div>
+      <div className="relative aspect-[5/3] w-full bg-[#2D2D2D]">이미지 1</div>
     </div>
   );
 };
