+++ src/lib/ImagesContext.jsx
@@ -1,0 +1,25 @@
+// lib/ImagesContext.jsx
+import { createContext, useContext, useState } from "react";
+
+const ImagesContext = createContext();
+
+export const ImagesProvider = ({ children }) => {
+  const [images, setImages] = useState([]);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState(null);
+
+  return (
+    <ImagesContext.Provider
+      value={{ images, loading, error, setImages, setLoading, setError }}
+    >
+      {children}
+    </ImagesContext.Provider>
+  );
+};
+
+export const useImages = () => {
+  const context = useContext(ImagesContext);
+  if (!context)
+    throw new Error("useImages must be used within an ImagesProvider");
+  return context;
+};
